+++ NEWS.md
@@ -1,3 +1,11 @@
+# RStoolbox 0.4.1
+## New:
+* `mesma()` now better differentiates SMA and MESMA: For single endmember unmixing, each supplied endmember represents a class to unmix (row by row). For multiple endmemeber unmixing, the column `class` can be used to group endmembers by class. If multiple endmembers per class are provided, `mesma()` will compute a number of SMA (determined through the new argument `n_models`) for multiple endmember combinations drawn from endmembers and will select the best fit per pixel based on the lowest RMSE. See `?mesma` for details (fixes #57, reported by @ytarazona)
+
+## Changes:
+* `mesma()` now implements the sum to one constraint by default (argument `sum_to_one`) (fixes #62, reported by @michaeldorman)
+* added a new example to `mesma()` to reflect the changes
+
 # RStoolbox 0.3.0
 ## New:
 * `rasterCVA()` by default no longer enforces a minimal change magnitude (can still be accomplished with the `tmf` argument).